collision/collision_checker/requirements.txt

dt-protocols-daffy<=6.2.24
zuper-nodes-z6>=6.2.2
shapely